backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
openai>=2.14.0
requests==2.31.0
